refactor: split Persona dialog
