+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Persona/index.tsx
@@ -1,0 +1,6 @@
+export { DashboardPersonaBackupDialog } from './Backup'
+export { DashboardPersonaCreateDialog } from './Create'
+export { DashboardPersonaDeleteConfirmDialog } from './DeleteConfirm'
+export { DashboardImportPersonaDialog } from './Import'
+export { DashboardPersonaRenameDialog } from './Rename'
+export { DashboardPersonaUnlinkConfirmDialog } from './UnlinkConfirm'
