+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Persona/types.ts
@@ -1,0 +1,5 @@
+import type { Persona } from '../../../../database'
+
+export interface PersonaProps {
+    persona: Persona
+}
